Backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0

# Authentication & Database
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
PyJWT==2.8.0
python-dotenv==1.0.0